Create method to provide a list of checking pieces
For checkmate, I needed a list of pieces putting the king in check
ir order to see if they could be captures. The previous method of
determing check was only true/false. The new approach created a new
method that can be called separetly when needed to determine checkmate.
At othertimes, the length of the array can be used to determine
if check exists at all.