Refactor to better identify where the king can move
Rather than trying to remove the king from the board, or remove
the piece on the finish square, for the purposes of looking for check
the king is now moved to that location. This better prevents "shadowing"
or other pieces, like pawns, from not knowing the squares they guard/
attack.