Update castling ability after corner capture
The rooks starting positions are the corners of the board. Since
there is already an update to the castling ability if the rook moves
this only looks for captures of the rook on their starting squares.
If the starting square is empty or a non-rook piece is there, the rook
has already moved and the castling ability won't be impacted.